+++ services/orchest-webserver/client/src/views/ProjectsView.tsx
@@ -14,6 +14,7 @@ import {
   RefManager,
   validURL,
 } from "@orchest/lib-utils";
+import type { TViewProps } from "@/types";
 import { useOrchest } from "@/hooks/orchest";
 import { BackgroundTaskPoller } from "@/utils/webserver-utils";
 import { Layout } from "@/components/Layout";
@@ -26,7 +27,7 @@ const ERROR_MAPPING = {
     "project name contains illegal character(s).",
 } as const;
 
-const ProjectsView: React.FC<any> = (props) => {
+const ProjectsView: React.FC<TViewProps> = (props) => {
   const { orchest } = window;
 
   const context = useOrchest();
@@ -34,6 +35,7 @@ const ProjectsView: React.FC<any> = (props) => {
   const [state, setState] = React.useState({
     createModal: false,
     showImportModal: false,
+    isDeleting: false,
     loading: true,
     // import dialog
     import_url: "",
@@ -44,6 +46,10 @@ const ProjectsView: React.FC<any> = (props) => {
     listData: null,
     importResult: undefined,
     fetchListAndSetProject: "",
+    editProjectPathModal: false,
+    editProjectPathModalBusy: false,
+    editProjectPathUUID: undefined,
+    editProjectPath: undefined,
   });
 
   const [promiseManager] = React.useState(new PromiseManager());
@@ -69,12 +75,85 @@ const ProjectsView: React.FC<any> = (props) => {
     }
   };
 
+  const onEditProjectName = (projectUUID, projectPath) => {
+    setState((prevState) => ({
+      ...prevState,
+      editProjectPathUUID: projectUUID,
+      editProjectPath: projectPath,
+      editProjectPathModal: true,
+    }));
+  };
+
+  const onCloseEditProjectPathModal = () => {
+    setState((prevState) => ({
+      ...prevState,
+      editProjectPathModal: false,
+      editProjectPathModalBusy: false,
+    }));
+  };
+
+  const onSubmitEditProjectPathModal = () => {
+    if (!validateProjectNameAndAlert(state.editProjectPath)) {
+      return;
+    }
+
+    setState((prevState) => ({
+      ...prevState,
+      editProjectPathModalBusy: true,
+    }));
+
+    makeRequest("PUT", `/async/projects/${state.editProjectPathUUID}`, {
+      type: "json",
+      content: {
+        name: state.editProjectPath,
+      },
+    })
+      .then((_) => {
+        fetchList();
+      })
+      .catch((e) => {
+        try {
+          let resp = JSON.parse(e.body);
+
+          if (resp.code == 0) {
+            orchest.alert(
+              "Error",
+              "Cannnot rename project when an interactive session is running."
+            );
+          } else if (resp.code == 1) {
+            orchest.alert(
+              "Error",
+              'Cannnot rename project, a project with the name "' +
+                state.editProjectPath +
+                '" already exists.'
+            );
+          }
+        } catch (error) {
+          console.error(e);
+          console.error(error);
+        }
+      })
+      .finally(() => {
+        onCloseEditProjectPathModal();
+      });
+  };
+
   const processListData = (projects) => {
     let listData = [];
 
     for (let project of projects) {
       listData.push([
-        <span>{project.path}</span>,
+        <span className="mdc-icon-table-wrapper">
+          {project.path}{" "}
+          <span className="consume-click">
+            <MDCIconButtonToggleReact
+              icon="edit"
+              onClick={() => {
+                onEditProjectName(project.uuid, project.path);
+              }}
+            />
+          </span>
+        </span>,
         <span>{project.pipeline_count}</span>,
         <span>{project.session_count}</span>,
         <span>{project.job_count}</span>,
@@ -145,34 +224,57 @@ const ProjectsView: React.FC<any> = (props) => {
   };
 
   const onDeleteClick = () => {
-    let selectedIndices = refManager.refs.projectListView.getSelectedRowIndices();
+    if (!state.isDeleting) {
+      setState((prevState) => ({
+        ...prevState,
+        isDeleting: true,
+      }));
 
-    if (selectedIndices.length === 0) {
-      orchest.alert("Error", "You haven't selected a project.");
-      return;
-    }
+      let selectedIndices = refManager.refs.projectListView.getSelectedRowIndices();
 
-    orchest.confirm(
-      "Warning",
-      "Are you certain that you want to delete this project? This will kill all associated resources and also delete all corresponding jobs. (This cannot be undone.)",
-      () => {
+      if (selectedIndices.length === 0) {
+        orchest.alert("Error", "You haven't selected a project.");
+
         setState((prevState) => ({
           ...prevState,
-          loading: true,
+          isDeleting: false,
         }));
 
-        let deletePromises = [];
-
-        selectedIndices.forEach((index) => {
-          let project_uuid = state.projects[index].uuid;
-          deletePromises.push(deleteProjectRequest(project_uuid));
-        });
-
-        Promise.all(deletePromises).then(() => {
-          fetchList();
-        });
+        return;
       }
-    );
+
+      orchest.confirm(
+        "Warning",
+        "Are you certain that you want to delete this project? This will kill all associated resources and also delete all corresponding jobs. (This cannot be undone.)",
+        () => {
+          // Start actual delete
+          let deletePromises = [];
+
+          selectedIndices.forEach((index) => {
+            let project_uuid = state.projects[index].uuid;
+            deletePromises.push(deleteProjectRequest(project_uuid));
+          });
+
+          Promise.all(deletePromises).then(() => {
+            fetchList();
+
+            // Clear isDeleting
+            setState((prevState) => ({
+              ...prevState,
+              isDeleting: false,
+            }));
+          });
+        },
+        () => {
+          setState((prevState) => ({
+            ...prevState,
+            isDeleting: false,
+          }));
+        }
+      );
+    } else {
+      console.error("Delete UI in progress.");
+    }
   };
 
   const deleteProjectRequest = (project_uuid) => {
@@ -213,18 +315,7 @@ const ProjectsView: React.FC<any> = (props) => {
   const onSubmitModal = () => {
     let projectName = state.create_project_name;
 
-    if (projectName.length == 0) {
-      orchest.alert("Error", "Project name cannot be empty.");
-      return;
-    }
-
-    let projectNameValidation = validProjectName(projectName);
-    if (!projectNameValidation.valid) {
-      orchest.alert(
-        "Error",
-        "Please make sure you enter a valid project name. " +
-          projectNameValidation.reason
-      );
+    if (!validateProjectNameAndAlert(projectName)) {
       return;
     }
 
@@ -261,7 +352,12 @@ const ProjectsView: React.FC<any> = (props) => {
   };
 
   const validProjectName = (projectName) => {
-    if (projectName.match("[^A-Za-z0-9_.-]")) {
+    if (projectName === undefined || projectName.length == 0) {
+      return {
+        valid: false,
+        reason: "Project name cannot be empty.",
+      };
+    } else if (projectName.match("[^A-Za-z0-9_.-]")) {
       return {
         valid: false,
         reason:
@@ -272,6 +368,18 @@ const ProjectsView: React.FC<any> = (props) => {
       };
     }
     return { valid: true };
+  };
+
+  const validateProjectNameAndAlert = (projectName) => {
+    let projectNameValidation = validProjectName(projectName);
+    if (!projectNameValidation.valid) {
+      orchest.alert(
+        "Error",
+        "Please make sure you enter a valid project name. " +
+          projectNameValidation.reason
+      );
+    }
+    return projectNameValidation.valid;
   };
 
   const onSubmitImport = () => {
@@ -286,13 +394,11 @@ const ProjectsView: React.FC<any> = (props) => {
       return;
     }
 
-    let projectNameValidation = validProjectName(gitProjectName);
-    if (!projectNameValidation.valid) {
-      orchest.alert(
-        "Error",
-        "Please make sure you enter a valid project name. " +
-          projectNameValidation.reason
-      );
+    // Note: can have empty project name for import (uses inferred repo name by git)
+    if (
+      gitProjectName.length > 0 &&
+      !validateProjectNameAndAlert(gitProjectName)
+    ) {
       return;
     }
 
@@ -408,100 +514,134 @@ const ProjectsView: React.FC<any> = (props) => {
   return (
     <Layout>
       <div className={"view-page projects-view"}>
-        {(() => {
-          if (state.showImportModal) {
-            return (
-              <MDCDialogReact
-                title="Import a project"
-                onClose={onCloseImportProjectModal.bind(this)}
-                ref={refManager.nrefs.importProjectDialog}
-                content={
-                  <div className="project-import-modal">
-                    {props.queryArgs &&
-                      props.queryArgs.import_url !== undefined && (
-                        <div className="push-down warning">
-                          <p>
-                            <i className="material-icons">warning</i> The import
-                            URL was pre-filled. Make sure you trust the{" "}
-                            <span className="code">git</span> repository you're
-                            importing.
-                          </p>
-                        </div>
-                      )}
-                    <p className="push-down">
-                      Import a <span className="code">git</span> repository by
-                      specifying the <span className="code">HTTPS</span> URL
-                      below:
-                    </p>
-                    <MDCTextFieldReact
-                      classNames={["fullwidth push-down"]}
-                      label="Git repository URL"
-                      value={state.import_url}
-                      onChange={handleChange.bind(this, "import_url")}
-                    />
-
-                    <MDCTextFieldReact
-                      classNames={["fullwidth"]}
-                      label="Project name (optional)"
-                      value={state.import_project_name}
-                      onChange={handleChange.bind(this, "import_project_name")}
-                    />
-
-                    {(() => {
-                      if (state.importResult) {
-                        let result;
-
-                        if (state.importResult.status === "PENDING") {
-                          result = <MDCLinearProgressReact />;
-                        } else if (state.importResult.status === "FAILURE") {
-                          result = (
-                            <p>
-                              <i className="material-icons float-left">error</i>{" "}
-                              Import failed:{" "}
-                              {getMappedErrorMessage(state.importResult.result)}
-                            </p>
-                          );
-                        }
-
-                        return <div className="push-up">{result}</div>;
-                      }
-                    })()}
-                    <p className="push-up">
-                      To import <b>private </b>
-                      <span className="code">git</span> repositories upload them
-                      directly through the File Manager into the{" "}
-                      <span className="code">projects/</span> directory.
+        {state.showImportModal && (
+          <MDCDialogReact
+            title="Import a project"
+            onClose={onCloseImportProjectModal.bind(this)}
+            ref={refManager.nrefs.importProjectDialog}
+            content={
+              <div className="project-import-modal">
+                {props.queryArgs && props.queryArgs.import_url !== undefined && (
+                  <div className="push-down warning">
+                    <p>
+                      <i className="material-icons">warning</i> The import URL
+                      was pre-filled. Make sure you trust the{" "}
+                      <span className="code">git</span> repository you're
+                      importing.
                     </p>
                   </div>
-                }
-                actions={
-                  <React.Fragment>
-                    <MDCButtonReact
-                      icon="close"
-                      label="Close"
-                      classNames={["push-right"]}
-                      onClick={onCancelImport.bind(this)}
-                    />
-                    <MDCButtonReact
-                      icon="input"
-                      // So that the button is disabled when in a states
-                      // that requires so (currently ["PENDING"]).
-                      disabled={["PENDING"].includes(
-                        state.importResult !== undefined
-                          ? state.importResult.status
-                          : undefined
-                      )}
-                      classNames={["mdc-button--raised", "themed-secondary"]}
-                      label="Import"
-                      submitButton
-                      onClick={onSubmitImport.bind(this)}
-                    />
-                  </React.Fragment>
-                }
-              />
-            );
-          }
-        })()}
+                )}
+                <p className="push-down">
+                  Import a <span className="code">git</span> repository by
+                  specifying the <span className="code">HTTPS</span> URL below:
+                </p>
+                <MDCTextFieldReact
+                  classNames={["fullwidth push-down"]}
+                  label="Git repository URL"
+                  value={state.import_url}
+                  onChange={handleChange.bind(this, "import_url")}
+                />
+
+                <MDCTextFieldReact
+                  classNames={["fullwidth"]}
+                  label="Project name (optional)"
+                  value={state.import_project_name}
+                  onChange={handleChange.bind(this, "import_project_name")}
+                />
+
+                {(() => {
+                  if (state.importResult) {
+                    let result;
+
+                    if (state.importResult.status === "PENDING") {
+                      result = <MDCLinearProgressReact />;
+                    } else if (state.importResult.status === "FAILURE") {
+                      result = (
+                        <p>
+                          <i className="material-icons float-left">error</i>{" "}
+                          Import failed:{" "}
+                          {getMappedErrorMessage(state.importResult.result)}
+                        </p>
+                      );
+                    }
+
+                    return <div className="push-up">{result}</div>;
+                  }
+                })()}
+                <p className="push-up">
+                  To import <b>private </b>
+                  <span className="code">git</span> repositories upload them
+                  directly through the File Manager into the{" "}
+                  <span className="code">projects/</span> directory.
+                </p>
+              </div>
+            }
+            actions={
+              <React.Fragment>
+                <MDCButtonReact
+                  icon="close"
+                  label="Close"
+                  classNames={["push-right"]}
+                  onClick={onCancelImport.bind(this)}
+                />
+                <MDCButtonReact
+                  icon="input"
+                  // So that the button is disabled when in a states
+                  // that requires so (currently ["PENDING"]).
+                  disabled={["PENDING"].includes(
+                    state.importResult !== undefined
+                      ? state.importResult.status
+                      : undefined
+                  )}
+                  classNames={["mdc-button--raised", "themed-secondary"]}
+                  label="Import"
+                  submitButton
+                  onClick={onSubmitImport.bind(this)}
+                />
+              </React.Fragment>
+            }
+          />
+        )}
+
+        {state.editProjectPathModal && (
+          <MDCDialogReact
+            title="Edit project name"
+            onClose={onCloseEditProjectPathModal}
+            content={
+              <React.Fragment>
+                <MDCTextFieldReact
+                  classNames={["fullwidth push-down"]}
+                  value={state.editProjectPath}
+                  label="Project name"
+                  onChange={(value) => {
+                    setState((prevState) => ({
+                      ...prevState,
+                      editProjectPath: value,
+                    }));
+                  }}
+                />
+              </React.Fragment>
+            }
+            actions={
+              <React.Fragment>
+                <MDCButtonReact
+                  icon="close"
+                  label="Cancel"
+                  classNames={["push-right"]}
+                  onClick={onCloseEditProjectPathModal}
+                />
+                <MDCButtonReact
+                  icon="save"
+                  disabled={state.editProjectPathModalBusy}
+                  classNames={["mdc-button--raised", "themed-secondary"]}
+                  label="Save"
+                  submitButton
+                  onClick={onSubmitEditProjectPathModal}
+                />
+              </React.Fragment>
+            }
+          />
+        )}
 
         {(() => {
           if (state.createModal) {
@@ -569,6 +709,7 @@ const ProjectsView: React.FC<any> = (props) => {
                 <div className={"pipeline-actions push-down"}>
                   <MDCIconButtonToggleReact
                     icon="delete"
+                    disabled={state.isDeleting}
                     tooltipText="Delete project"
                     onClick={onDeleteClick.bind(this)}
                   />
